+++ AutoStack/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.0.3]
+- added: resource chests output instantly and limit to 2 units per type (off by default)
+- unlocked view modes (energy, sewers)
+
 ## [1.0.2]
 - added: resources stack on resource camps